+++ kotlin-react/src/main/kotlin/react/UseLayoutEffect.kt
@@ -1,0 +1,35 @@
+package react
+
+/**
+ * Only works inside [functionalComponent]
+ * @see <a href="https://reactjs.org/docs/hooks-state.html#hooks-and-function-components">Hooks and Function Components</a>
+ */
+fun useLayoutEffectWithCleanup(
+    dependencies: RDependenciesList? = null,
+    effect: () -> RCleanup,
+) {
+    if (dependencies != null) {
+        rawUseLayoutEffect(effect, dependencies.toTypedArray())
+    } else {
+        rawUseLayoutEffect(effect)
+    }
+}
+
+/**
+ * Only works inside [functionalComponent]
+ * @see <a href="https://reactjs.org/docs/hooks-state.html#hooks-and-function-components">Hooks and Function Components</a>
+ */
+fun useLayoutEffect(
+    dependencies: RDependenciesList? = null,
+    effect: () -> Unit,
+) {
+    val rawEffect = {
+        effect()
+        undefined
+    }
+    if (dependencies != null) {
+        rawUseLayoutEffect(rawEffect, dependencies.toTypedArray())
+    } else {
+        rawUseLayoutEffect(rawEffect)
+    }
+}
